ai-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pdfplumber
sentence-transformers
scikit-learn
torch
python-dotenv
requests
